+++ java/edu/ualr/bittorrent/interfaces/TrackerResponse.java
@@ -5,10 +5,10 @@ import com.google.common.collect.ImmutableList;
 public interface TrackerResponse {
   public String getFailureReason();
   public String getWarningMessage();
-  public Integer getInterval();
+  public int getInterval();
   public Integer getMinInterval();
   public byte[] getTrackerId();
-  public Long getComplete();
-  public Long getIncomplete();
+  public int getComplete();
+  public int getIncomplete();
   public ImmutableList<Peer> getPeers();
 }
